Moved utils from model to dao.
