+++ src/test/java/pl/ciesielski/dominik/app/cardealerapp/dao/ClientDaoIntegrationTest.java
@@ -11,7 +11,7 @@ import pl.ciesielski.dominik.app.cardealerapp.dao.utils.DatabaseInitializer;
 import pl.ciesielski.dominik.app.cardealerapp.model.Address;
 import pl.ciesielski.dominik.app.cardealerapp.model.AddressBuilder;
 import pl.ciesielski.dominik.app.cardealerapp.model.Client;
-import pl.ciesielski.dominik.app.cardealerapp.model.utils.NextIdSequence;
+import pl.ciesielski.dominik.app.cardealerapp.dao.utils.NextIdSequence;
 
 import java.sql.Connection;
 import java.sql.SQLException;
