+++ src/test/java/pl/ciesielski/dominik/app/cardealerapp/dao/ClientDaoTest.java
@@ -6,7 +6,7 @@ import pl.ciesielski.dominik.app.cardealerapp.dao.utils.DatabaseInitializer;
 import pl.ciesielski.dominik.app.cardealerapp.model.Address;
 import pl.ciesielski.dominik.app.cardealerapp.model.AddressBuilder;
 import pl.ciesielski.dominik.app.cardealerapp.model.Client;
-import pl.ciesielski.dominik.app.cardealerapp.model.utils.NextIdSequence;
+import pl.ciesielski.dominik.app.cardealerapp.dao.utils.NextIdSequence;
 
 import java.sql.Connection;
 import java.sql.SQLException;
